requirements/base.txt
Django==4.1
django-environ==0.9.0
environ==1.0
Pillow==9.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-health-check==3.16.5
django-cors-headers==3.13.0
drf-yasg==1.21.3
celery==5.2.7
Redis==4.3.4
django_celery_results==2.4.0

psutil
dj-database-url
uuid
docx2txt
requests-async